added rudimentary tmat test
