+++ simpletest/test.py
@@ -1,56 +1,30 @@
 #!/usr/bin/env python
-import os
-from os import path
+from os import path, makedirs
 import p2ptrans as p2p
 from pylada.crystal import read
-import itertools
-import numpy as np
+from test_utils import *
+import pickle
 
 bcc = read.poscar(path.join('ref','BCC_Fe_POSCAR'))
 fcc = read.poscar(path.join('ref','FCC_Fe_POSCAR'))
 
 outdir = 'results'
-if not os.path.exists(outdir):
-   os.makedirs(outdir) 
+if not path.exists(outdir):
+   makedirs(outdir) 
 
 # n_steps, (tmat, dispStruc, vec_classes), outdir, display
 tmat, dispStruc, vec_classes = p2p.findMatching(bcc, fcc, 100, outdir=outdir)
 p2p.produceTransition(60, tmat, dispStruc, vec_classes, outdir, False)
 
+# check that tmats are close
+pickle.dump( tmat, open(path.join(outdir,'tmat.pickle'), 'wb') )
+ref_tmat = pickle.load(open(path.join('ref','tmat.pickle'), 'rb'))
+if not compare_tmats(ref_tmat, tmat, tol=1e-4):
+    print('FAILED TMATS!!!!')
+    exit()
 
 
-def poscars_in_dir(main_dir):
-    poscars = []
-    for i in itertools.count():
-        poscar_path = path.join(main_dir, 'POSCAR_{:03}'.format(i))
-        if path.exists(poscar_path):
-            poscars.append(poscar_path)
-        else:
-            return poscars
-
-def structs_in_dir(main_dir):
-    return [read.poscar(poscar) for poscar in poscars_in_dir(main_dir)]
-
-def compare_structs(struct_a, struct_b, tol=1e-4):
-    '''Will produce false positives (cell) and false negitives (supercells, transposed atoms)'''
-    cells = compare_cells(struct_a, struct_b)
-    num_atoms = len(struct_a) == len(struct_b)
-    atom_poses = all([np.allclose(a.pos, b.pos, atol=tol) for a, b in zip(struct_a, struct_b)])
-    atom_types = all([a.type == b.type for a, b in zip(struct_a, struct_b)])
-    return cells and num_atoms and atom_poses and atom_types
-    
-def compare_cells(struct_a, struct_b, tol=1e-4):
-    '''Will produce false positives (doesn't check angles)'''
-    cell_vols = np.isclose(struct_a.volume, struct_b.volume, atol=tol)
-    cell_vectors_a_lens = np.linalg.norm(struct_a.cell, axis=0)
-    cell_vectors_b_lens = np.linalg.norm(struct_b.cell, axis=0)
-    cell_vectors_a_lens = np.sort(cell_vectors_a_lens)
-    cell_vectors_b_lens = np.sort(cell_vectors_b_lens)
-    return np.allclose(cell_vectors_a_lens, cell_vectors_b_lens, atol=tol) and cell_vols 
-
-
-
-# then verifiy
+# then verifiy that produced poscars are close to reference
 ref_structs = structs_in_dir(path.join('ref', 'TransPOSCARS'))
 produced_structs = structs_in_dir(path.join('results', 'TransPOSCARS'))
 
@@ -62,7 +36,6 @@ for ref_struct, res_struct in zip(ref_structs, produced_structs):
         exit()
 
 print('Succeded!')
-
